Modificar valor de retorno

+++ include/stats.h
@@ -2,7 +2,7 @@
 #define STATS_H
 
 int   suma    (int *arr, int n);
-int   promedio(int *arr, int n);
+float   promedio(int *arr, int n);
 int   minimo  (int *arr, int n);
 int   maximo  (int *arr, int n);
